# src/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "gh-proxy",
    "description": "A community-run Hack Club GitHub API proxy. Anyone with an eligible Hack Club Auth account can get a rate-limited API key by donating a GitHub token; anyone at all can donate a token to the shared pool.",
    "version": "2.0.0"
  },
  "paths": {
...
    "/admin": { "get": { "summary": "Admin dashboard (Hack Club Auth + allow-list)", "responses": { "200": { "description": "OK" }, "403": { "$ref": "#/components/responses/Forbidden" } } } },
...
    "/admin/apikeys/{id}/disable": { "post": { "summary": "Disable an API key (admin only)", "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }], "responses": { "303": { "description": "Redirect back to /admin" } } } }
  },
  "components": {
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": { "type": "string" },
              "message": { "type": "string" },
              "hint": { "type": "string" },
              "documentation_url": { "type": "string" },
              "links": { "type": "array", "items": { "type": "object", "properties": { "rel": { "type": "string" }, "href": { "type": "string" }, "description": { "type": "string" } } } }
            },
            "required": ["code", "message"]
          }
        }
      }
    },
    "responses": {
...
      "Forbidden": { "description": "Key disabled, or caller not on the admin allow-list", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
...
    },
    "securitySchemes": {
      "ApiKeyAuth": { "type": "apiKey", "in": "header", "name": "X-API-Key" }
    }
  },
  "security": [{ "ApiKeyAuth": [] }]
}